fixing realtime progress

+++ src/main/java/com/fajar/imagemosaic/tools/MosaicProcessNotifier.java
@@ -2,5 +2,5 @@ package com.fajar.imagemosaic.tools;
 
 public interface MosaicProcessNotifier {
 
-	void notify(int step, int totalStep);
+	void notify(Double step, Double totalStep);
 }
